Create a proper Chosen class
This way the event to wait for and person are in the same place,
which hopefully should avoid issues with clearing the event because
now it's brand new every time.

In addition we stop monkey-patching the User to have a .name field
and the type of the chosen global becomes clear.